Add files via upload
Created a program with a concurrent server using TCP that can handle multiple requests to sort an array of twenty randomly generated words using mergesort that are submitted to the server by each of the clients. The arrays are created and stored on the clients in a separate input file called "input.dat", where n is the client number before starting the program itself.